refactor: remove Configuration as a base of MirroringConfiguration (#127)
