+++ bigtable-hbase-mirroring-client-1.x-parent/bigtable-hbase-mirroring-client-1.x/src/main/java/com/google/cloud/bigtable/mirroring/hbase1_x/bufferedmutator/MirroringBufferedMutator.java
@@ -156,7 +156,7 @@ public abstract class MirroringBufferedMutator<BufferEntryType> implements Buffe
     this.mutationsBufferFlushIntervalBytes =
         configuration.mirroringOptions.bufferedMutatorBytesToFlush;
     this.executorService = MoreExecutors.listeningDecorator(executorService);
-    this.configuration = configuration;
+    this.configuration = configuration.baseConfiguration;
     this.bufferedMutatorParams = bufferedMutatorParams;
 
     this.mutationEntries = new ArrayList<>();
