+++ bigtable-hbase-mirroring-client-1.x-parent/bigtable-hbase-mirroring-client-1.x/src/test/java/com/google/cloud/bigtable/mirroring/hbase1_x/TestConnection.java
@@ -1,0 +1,98 @@
+/*
+ * Copyright 2021 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.google.cloud.bigtable.mirroring.hbase1_x;
+
+import static org.mockito.Mockito.mock;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.concurrent.ExecutorService;
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.hbase.TableName;
+import org.apache.hadoop.hbase.client.Admin;
+import org.apache.hadoop.hbase.client.BufferedMutator;
+import org.apache.hadoop.hbase.client.BufferedMutatorParams;
+import org.apache.hadoop.hbase.client.Connection;
+import org.apache.hadoop.hbase.client.RegionLocator;
+import org.apache.hadoop.hbase.client.Table;
+import org.apache.hadoop.hbase.security.User;
+
+public class TestConnection implements Connection {
+  public static List<Connection> mocks = new ArrayList<>();
+  private Connection connectionMock;
+
+  public TestConnection(Configuration conf, boolean managed, ExecutorService pool, User user) {
+    connectionMock = mock(Connection.class);
+    mocks.add(connectionMock);
+  }
+
+  @Override
+  public Configuration getConfiguration() {
+    return connectionMock.getConfiguration();
+  }
+
+  @Override
+  public Table getTable(TableName tableName) throws IOException {
+    return connectionMock.getTable(tableName);
+  }
+
+  @Override
+  public Table getTable(TableName tableName, ExecutorService executorService) throws IOException {
+    return connectionMock.getTable(tableName, executorService);
+  }
+
+  @Override
+  public BufferedMutator getBufferedMutator(TableName tableName) throws IOException {
+    return connectionMock.getBufferedMutator(tableName);
+  }
+
+  @Override
+  public BufferedMutator getBufferedMutator(BufferedMutatorParams bufferedMutatorParams)
+      throws IOException {
+    return connectionMock.getBufferedMutator(bufferedMutatorParams);
+  }
+
+  @Override
+  public RegionLocator getRegionLocator(TableName tableName) throws IOException {
+    return connectionMock.getRegionLocator(tableName);
+  }
+
+  @Override
+  public Admin getAdmin() throws IOException {
+    return connectionMock.getAdmin();
+  }
+
+  @Override
+  public void close() throws IOException {
+    connectionMock.close();
+  }
+
+  @Override
+  public boolean isClosed() {
+    return connectionMock.isClosed();
+  }
+
+  @Override
+  public void abort(String s, Throwable throwable) {
+    connectionMock.abort(s, throwable);
+  }
+
+  @Override
+  public boolean isAborted() {
+    return connectionMock.isAborted();
+  }
+}
