+++ bigtable-hbase-mirroring-client-1.x-parent/bigtable-hbase-mirroring-client-1.x/src/test/java/com/google/cloud/bigtable/mirroring/hbase1_x/TestMirroringConnection.java
@@ -20,93 +20,16 @@ import static com.google.cloud.bigtable.mirroring.hbase1_x.utils.MirroringConfig
 import static com.google.cloud.bigtable.mirroring.hbase1_x.utils.MirroringConfigurationHelper.MIRRORING_SECONDARY_CONFIG_PREFIX_KEY;
 import static com.google.cloud.bigtable.mirroring.hbase1_x.utils.MirroringConfigurationHelper.MIRRORING_SECONDARY_CONNECTION_CLASS_KEY;
 import static com.google.common.truth.Truth.assertThat;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.concurrent.ExecutorService;
 import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.hbase.TableName;
-import org.apache.hadoop.hbase.client.Admin;
-import org.apache.hadoop.hbase.client.BufferedMutator;
-import org.apache.hadoop.hbase.client.BufferedMutatorParams;
 import org.apache.hadoop.hbase.client.Connection;
 import org.apache.hadoop.hbase.client.ConnectionFactory;
-import org.apache.hadoop.hbase.client.RegionLocator;
-import org.apache.hadoop.hbase.client.Table;
-import org.apache.hadoop.hbase.security.User;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
-
-class TestConnection implements Connection {
-  public static List<Connection> mocks = new ArrayList<>();
-  private Connection connectionMock;
-
-  public TestConnection(Configuration conf, boolean managed, ExecutorService pool, User user) {
-    connectionMock = mock(Connection.class);
-    mocks.add(connectionMock);
-  }
-
-  @Override
-  public Configuration getConfiguration() {
-    return connectionMock.getConfiguration();
-  }
-
-  @Override
-  public Table getTable(TableName tableName) throws IOException {
-    return connectionMock.getTable(tableName);
-  }
-
-  @Override
-  public Table getTable(TableName tableName, ExecutorService executorService) throws IOException {
-    return connectionMock.getTable(tableName, executorService);
-  }
-
-  @Override
-  public BufferedMutator getBufferedMutator(TableName tableName) throws IOException {
-    return connectionMock.getBufferedMutator(tableName);
-  }
-
-  @Override
-  public BufferedMutator getBufferedMutator(BufferedMutatorParams bufferedMutatorParams)
-      throws IOException {
-    return connectionMock.getBufferedMutator(bufferedMutatorParams);
-  }
-
-  @Override
-  public RegionLocator getRegionLocator(TableName tableName) throws IOException {
-    return connectionMock.getRegionLocator(tableName);
-  }
-
-  @Override
-  public Admin getAdmin() throws IOException {
-    return connectionMock.getAdmin();
-  }
-
-  @Override
-  public void close() throws IOException {
-    connectionMock.close();
-  }
-
-  @Override
-  public boolean isClosed() {
-    return connectionMock.isClosed();
-  }
-
-  @Override
-  public void abort(String s, Throwable throwable) {
-    connectionMock.abort(s, throwable);
-  }
-
-  @Override
-  public boolean isAborted() {
-    return connectionMock.isAborted();
-  }
-}
 
 @RunWith(JUnit4.class)
 public class TestMirroringConnection {
